pyproject.toml
[project]
name = "psych"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyglet>=2.1.6",
    "pyyaml>=6.0.2",
]

[project.scripts]
test = "main"